requirements.txt
fastapi
uvicorn
openai
PyMuPDF
python-dotenv
pydantic
requests
